BUGFIX:
The module code which gets added as a tag to each Exhibition needs the proper conversion from the STePs DB to our DB, which enforces a lowercase constraint to each tag. If this is not done, a duplicate of each module code gets inserted. See Line 146 -> 152 for the BUGFIX.

Also added a helper function which removes duplicates from a given Array.